// ui/src/generated/CollabStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SiteId } from "./SiteId";

export type CollabStatus = { "type": "Offline" } | { "type": "Connecting", session: string, } | { "type": "Online", session: string, site: SiteId, };